requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.8.3
click==8.3.0
colorama==0.4.6
databases==0.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.13
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.3
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.3.4
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.3
rich==14.1.0
rich-toolkit==0.15.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.19
starlette==0.37.2
typer==0.19.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
apscheduler==3.10.4
aiokafka==0.10.0
aiosmtplib==5.0.0
